requirements.txt
uvicorn
fastapi
pydantic
joblib
scikit-learn
xgboost
pandas
requests
beautifulsoup4
pydantic[email]
tortoise-orm
aiomysql
passlib[bcrypt]
pydantic-settings
pyjwt[crypto]
google-generativeai
pytest
pytest-asyncio
python-multipart
selenium
webdriver-manager
undetected_chromedriver
aerich